requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
boto3==1.34.0
pydub==0.25.1
moviepy==1.0.3
numpy==1.24.3
openai==1.3.0
sentence-transformers==2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.1
Pillow==10.1.0
requests==2.31.0